// FRONT/financial/src/types/generated/aimDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0
 */
import type { AimProgressDto } from './aimProgressDto';
import type { CurrencyDto } from './currencyDto';
import type { SourceDtoLookup } from './sourceDtoLookup';

export interface AimDto {
  id?: number;
  /** @nullable */
  name?: string | null;
  amount?: number;
  priority?: number;
  userId?: number;
  isClosed?: boolean;
  currency?: CurrencyDto;
  /** @nullable */
  sources?: SourceDtoLookup[] | null;
  progress?: AimProgressDto;
}